Task-134 Syncrpchandler: change handler queueName
