+++ pkg/syncrpc/responses_consumer.go
@@ -52,19 +52,19 @@ func (h *Handler) responsesConsumerErrorCallback(_ *rmqworker.RMQWorker, err *co
 
 func (h *Handler) startResponsesConsumer(props HandlerProps) error {
 	spec := rmqworker.WorkerTask{
-		QueueName:                  h.queueName,
-		ISQueueDurable:             true,
-		ISAutoDelete:               false,
-		Callback:                   h.responsesConsumerCallback,
-		FromExchange:               props.ResponsesExchange,
-		ExchangeType:               exchangeType,
-		ConsumersCount:             consumersCount,
-		WorkerName:                 "syncrpc:" + h.props.ServiceTag,
-		MessagesLifetime:           messagesLifetimeMS,
-		RoutingKey:                 h.queueName,
-		UseErrorCallback:           true,
-		ErrorCallback:              h.responsesConsumerErrorCallback,
-		DisableCheckResponseErrors: true,
+		QueueName:        h.queueName,
+		ISQueueDurable:   true,
+		ISAutoDelete:     false,
+		Callback:         h.responsesConsumerCallback,
+		FromExchange:     props.ResponsesExchange,
+		ExchangeType:     exchangeType,
+		ConsumersCount:   consumersCount,
+		WorkerName:       "worker:" + h.queueName,
+		MessagesLifetime: messagesLifetimeMS,
+		RoutingKey:       h.queueName,
+		UseErrorCallback: true,
+		ErrorCallback:    h.responsesConsumerErrorCallback,
+		//DisableCheckResponseErrors: true,
 		//QueueLength:                props.QueueLength,
 		//DisableOverflow:            props.DisableOverflow,
 	}
